Restart logging after deploying
There's a bug somewhere in the Elastic Beanstalk stack which sometimes
results in the awslogs agent not streaming the new stdouterr file
when the docker container cycles during a deploy.

Restarting the awslogs service alone does not seem to be enough to
get things going again so we reset the state too which seems to work.